# Documents/ApiSpec/openapi.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "API",
    "description": "API Spec f�r den digitalen Reisef�hrer",
    "version": "v1"
  },
  "paths": {
...
    "/api/pointOfInterest": {
      "post": {
        "tags": [
          "PointOfInterest"
        ],
        "summary": "Add a new poi to the database",
        "operationId": "AddPOI",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/PostPointOfInterest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostPointOfInterest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostPointOfInterest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostPointOfInterest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PointOfInterest"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "PointOfInterest"
        ],
        "summary": "Edits the poi to a given id",
        "operationId": "PutPOI",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/PutPointOfInterest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutPointOfInterest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PutPointOfInterest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PutPointOfInterest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/pointOfInterest/{poiID}": {
      "delete": {
        "tags": [
          "PointOfInterest"
        ],
        "summary": "Deletes the poi to a given id",
        "operationId": "DeletePOI",
        "parameters": [
          {
            "name": "poiID",
            "in": "path",
            "description": "the poiID to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "PointOfInterest"
        ],
        "summary": "Gets the poi to a given id",
        "operationId": "GetPOI",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "poiID",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PointOfInterest"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/pointOfInterest/{userName}/center": {
      "get": {
        "tags": [
          "PointOfInterest"
        ],
        "summary": "Get the center of all poi's from the given user",
        "operationId": "GetCenterOfPOIsAsync",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userName",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CenterResult"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/pointOfInterest/getUserPoIs/{userName}": {
      "get": {
        "tags": [
          "PointOfInterest"
        ],
        "summary": "Get all poi's from the given user",
        "operationId": "GetPOIs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userName",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PointOfInterest"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/pointOfInterest/getUserPoIs/all": {
      "get": {
        "tags": [
          "PointOfInterest"
        ],
        "summary": "Get all poi's",
        "operationId": "GetAllPOIs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PointOfInterest"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/pointOfInterest/promoted": {
      "get": {
        "tags": [
          "PointOfInterest"
        ],
        "summary": "Get all promoted pois",
        "operationId": "GetPromotedPoIs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PointOfInterest"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "CenterResult": {
        "required": [
          "latitude",
          "longitude"
        ],
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "description": "Gets or Sets Latitude",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "description": "Gets or Sets Longitude",
            "format": "double"
          }
        },
        "additionalProperties": false,
        "description": "Model class for a centerResult"
      },
...
      "PointOfInterest": {
        "required": [
          "averageRating",
          "description",
          "imageUrl",
          "isPromoted",
          "latitude",
          "longitude",
          "name",
          "numberOfRatings",
          "poIID",
          "userName"
        ],
        "type": "object",
        "properties": {
          "poIID": {
            "type": "string",
            "description": "Gets or Sets Id",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Gets or Sets Name"
          },
          "userName": {
            "type": "string",
            "description": "Gets or Sets username"
          },
          "description": {
            "type": "string",
            "description": "Gets or Sets Description"
          },
          "latitude": {
            "type": "number",
            "description": "Gets or Sets Latitude",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "description": "Gets or Sets Longitude",
            "format": "double"
          },
          "averageRating": {
            "type": "number",
            "description": "Gets or Sets AverageRating",
            "format": "double"
          },
          "numberOfRatings": {
            "type": "integer",
            "description": "Gets or Sets NumberOfRatings",
            "format": "int32"
          },
          "imageUrl": {
            "type": "string",
            "description": "Gets or Sets ImageUrl"
          },
          "isPromoted": {
            "type": "boolean",
            "description": "Defines, whether this poi is promoted or not"
          }
        },
        "additionalProperties": false,
        "description": "Model class for a PointOfInterest"
      },
...
      "PostPointOfInterest": {
        "required": [
          "description",
          "imageUrl",
          "latitude",
          "longitude",
          "name",
          "userName"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Gets or Sets Name"
          },
          "userName": {
            "type": "string",
            "description": "Gets or Sets username"
          },
          "description": {
            "type": "string",
            "description": "Gets or Sets Description"
          },
          "latitude": {
            "type": "number",
            "description": "Gets or Sets Latitude",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "description": "Gets or Sets Longitude",
            "format": "double"
          },
          "imageUrl": {
            "type": "string",
            "description": "Gets or Sets ImageUrl"
          }
        },
        "additionalProperties": false,
        "description": "Model class for posting a new poi"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "PutPointOfInterest": {
        "required": [
          "description",
          "id",
          "imageUrl",
          "latitude",
          "longitude",
          "name",
          "userName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets or Sets Id"
          },
          "name": {
            "type": "string",
            "description": "Gets or Sets Name"
          },
          "userName": {
            "type": "string",
            "description": "Gets or Sets username"
          },
          "description": {
            "type": "string",
            "description": "Gets or Sets Description"
          },
          "latitude": {
            "type": "number",
            "description": "Gets or Sets Latitude",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "description": "Gets or Sets Longitude",
            "format": "double"
          },
          "imageUrl": {
            "type": "string",
            "description": "Gets or Sets ImageUrl"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}