# WebApp/tourguide/src/app/api/model/centerResult.ts
/**
 * API
 * API Spec f�r den digitalen Reisef�hrer
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Model class for a centerResult
 */
export interface CenterResult { 
    /**
     * Gets or Sets Latitude
     */
    latitude: number;
    /**
     * Gets or Sets Longitude
     */
    longitude: number;
}